Move options out of game and into the main
